server/requirements.txt
Flask
gunicorn
google-genai
requests # <--- CRÍTICO: Asegúrate de que 'requests' esté aquí para el validador